backend/pyproject.toml
[project]
name = "ai-knowledge-assistant"
version = "0.1.0"
requires-python = ">=3.13,<4.0"

[tool.poetry]
packages = [
    { include = "app" },
]   

[tool.poetry.dependencies]
fastapi = "^0.115.0"
uvicorn = {extras = ["standard"], version = "^0.31.0"}
sentence-transformers = "^3.0.1"
faiss-cpu = "^1.8.0"
PyPDF2 = "^3.0.1"
python-multipart = "^0.0.9"
requests = "^2.32.3"

[tool.poetry.scripts]
server = "app.main:start"
